refactor: config system
Oh lord. If this PR is not being merged before crucial code changes are made, this PR will turn into a huge merge conflict. This refactor is huge and usually would deserve its own PR but since this theme system PR also contains a lot of changes and relies on the refactor, it has to be combined.